+++ src/main/java/com/revature/aspects/Verify.java
@@ -1,0 +1,8 @@
+package com.revature.aspects;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+public @interface Verify {
+}
